+++ Gruntfile.js
@@ -1,24 +1,24 @@
 module.exports = function (grunt) {
-	var staticExampleFiles = [ 'src/**/example/**', '!src/**/example/**/*.js' ];
+	var staticExampleFiles = [ '*/example/**', '!*/example/**/*.js' ];
 
 	require('grunt-dojo2').initConfig(grunt, {
 		copy: {
 			staticExampleFiles: {
 				expand: true,
-				cwd: '.',
+				cwd: 'src',
 				src: staticExampleFiles,
 				dest: '<%= devDirectory %>'
 			},
 			devStyles: {
 				expand: true,
-				cwd: '.',
-				src: 'src/styles/widgets.css',
+				cwd: 'src',
+				src: 'common/styles/widgets.css',
 				dest: '<%= devDirectory %>'
 			},
 			distStyles: {
 				expand: true,
 				cwd: 'src',
-				src: 'styles/widgets.css',
+				src: 'common/styles/widgets.css',
 				dest: '<%= distDirectory %>'
 			}
 		}
