+++ src/common/styles/animations.css
