+++ src/common/styles/animations.css.d.ts
