+++ src/common/styles/base.css
@@ -13,13 +13,13 @@
 /* Allows visually hidden, focusable content to become visible when navigated to via keyboard */
 .visuallyHidden.focusable:active,
 .visuallyHidden.focusable:focus {
-    clip: auto;
-    height: auto;
-    margin: 0;
-    overflow: visible;
-    position: static;
-    width: auto;
-    white-space: inherit;
+	clip: auto;
+	height: auto;
+	margin: 0;
+	overflow: visible;
+	position: static;
+	width: auto;
+	white-space: inherit;
 }
 
 /* If you're using .hidden, you probably mean it. */
