+++ src/common/styles/widgets.css
@@ -1,0 +1,4 @@
+@import 'animations.css';
+@import 'base.css';
+@import '../../dialog/styles/dialog.css';
+@import '../../slidepane/styles/slidePane.css';
