+++ src/common/tests/intern.ts
@@ -61,10 +61,10 @@ export const loaderOptions = {
 };
 
 // Non-functional test suite(s) to run in each browser
-export const suites = [ 'tests/unit/all' ];
+export const suites = [ 'src/common/tests/unit/all' ];
 
 // Functional test suite(s) to run in each browser once non-functional tests are completed
-export const functionalSuites = [ 'tests/functional/all' ];
+export const functionalSuites = [];
 
 // A regular expression matching URLs to files that should not be included in code coverage analysis
 export const excludeInstrumentation = /(?:node_modules|bower_components|tests|styles|example|.+Element\.)/;
