+++ src/common/tests/run.html
@@ -3,7 +3,7 @@
 <head>
 	<meta charset="utf-8">
 	<title>Intern suite</title>
-	<meta http-equiv="refresh" content="0;url=../node_modules/intern/client.html?config=_build/tests/intern">
+	<meta http-equiv="refresh" content="0;url=../../../../node_modules/intern/client.html?config=_build/tests/intern">
 </head>
 <body>
 	Redirecting to Intern client
