+++ src/common/tests/unit/all.ts
@@ -1,0 +1,4 @@
+import './main';
+import '../../../dialog/tests/unit/Dialog';
+import '../../../slidepane/tests/unit/SlidePane';
+import '../../../label/tests/unit/Label';
