+++ src/common/tests/unit/main.ts
@@ -1,6 +1,6 @@
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
-import * as widgets from '../../src/main';
+import * as widgets from '../../../main';
 
 registerSuite({
 	name: 'main',
