+++ src/dialog/example/index.html
@@ -6,7 +6,7 @@
 <body>
 	<script src="../../../../node_modules/@dojo/loader/loader.min.js"></script>
 	<script src="https://cdnjs.cloudflare.com/ajax/libs/maquette/2.4.2/css-transitions.min.js"></script>
-	<link rel="stylesheet" href="../../styles/widgets.css">
+	<link rel="stylesheet" href="../../common/styles/widgets.css">
 	<style>
 		.slideIn {
 			transition: top .2s;
