+++ src/dialog/styles/dialog.css
@@ -1,4 +1,4 @@
-@import '../../styles/variables.css';
+@import '../../common/styles/variables.css';
 
 .main {
 	position: absolute;
