+++ src/dialog/tests/unit/Dialog.ts
@@ -1,7 +1,7 @@
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
 import { VNode } from '@dojo/interfaces/vdom';
-import Dialog, { Role } from '../../src/dialog/Dialog';
+import Dialog, { Role } from '../../Dialog';
 
 registerSuite({
 	name: 'Dialog',
