+++ src/slidepane/styles/slidePane.css
@@ -1,4 +1,4 @@
-@import '../../styles/variables.css';
+@import '../../common/styles/variables.css';
 
 .underlay {
 	position: fixed;
