+++ src/slidepane/tests/unit/SlidePane.ts
@@ -1,8 +1,8 @@
 import * as registerSuite from 'intern!object';
 import * as assert from 'intern/chai!assert';
 import { VNode } from '@dojo/interfaces/vdom';
-import SlidePane, { Align } from '../../src/slidepane/SlidePane';
-import * as css from '../../src/slidepane/styles/slidePane.css';
+import SlidePane, { Align } from '../../SlidePane';
+import * as css from '../../styles/slidePane.css';
 
 function createEvent(type: string, x: number): any {
 	return {
