--- src/styles/widgets.css
@@ -1,4 +1,0 @@
-@import 'animations.css';
-@import 'base.css';
-@import '../dialog/styles/dialog.css';
-@import '../slidepane/styles/slidePane.css';
