--- tests/functional/all.ts
@@ -1,1 +1,0 @@
-export const removeThis = 1;
