--- tests/support/util.ts
@@ -1,18 +1,0 @@
-/**
- * Thenable represents any object with a callable `then` property.
- */
-export interface Thenable<T> {
-	then<U>(onFulfilled?: (value?: T) => U | Thenable<U>, onRejected?: (error?: any) => U | Thenable<U>): Thenable<U>;
-}
-
-export function isEventuallyRejected<T>(promise: Thenable<T>): Thenable<boolean> {
-	return promise.then<any>(function () {
-		throw new Error('unexpected code path');
-	}, function () {
-		return true; // expect rejection
-	});
-}
-
-export function throwImmediatly() {
-	throw new Error('unexpected code path');
-}
