--- tests/unit/all.ts
@@ -1,4 +1,0 @@
-import './main';
-import './Dialog';
-import './SlidePane';
-import './Label';
