+++ tsconfig.json
@@ -8,7 +8,7 @@
 		"moduleResolution": "node",
 		"noImplicitAny": true,
 		"noImplicitThis": true,
-		"outDir": "_build/",
+		"outDir": "_build/src/",
 		"removeComments": false,
 		"sourceMap": true,
 		"strictNullChecks": true,
@@ -17,7 +17,6 @@
 	},
 	"include": [
 		"./src/**/*.ts",
-		"./tests/**/*.ts",
 		"./typings/index.d.ts"
 	]
 }
